+++ hug/route.py
@@ -42,11 +42,14 @@ class Object(http):
     def __init__(self, urls=None, accept=HTTP_METHODS, output=None, **kwargs):
         super().__init__(urls=urls, accept=accept, output=output, **kwargs)
 
-    def __call__(self, method_or_class):
+    def __call__(self, method_or_class=None, **kwargs):
+        if not method_or_class and kwargs:
+            return self.where(**kwargs)
+
         if isinstance(method_or_class, (MethodType, FunctionType)):
-            routes = getattr(method_or_class, '_hug_routes', [])
+            routes = getattr(method_or_class, '_hug_http_routes', [])
             routes.append(self.route)
-            method_or_class._hug_routes = routes
+            method_or_class._hug_http_routes = routes
             return method_or_class
 
         instance = method_or_class
@@ -55,10 +58,14 @@ class Object(http):
 
         for argument in dir(instance):
             argument = getattr(instance, argument, None)
-            routes = getattr(argument, '_hug_routes', None)
-            if routes:
-                for route in routes:
-                    http(**self.where(**route).route)(argument)
+
+            http_routes = getattr(argument, '_hug_http_routes', ())
+            for route in http_routes:
+                http(**self.where(**route).route)(argument)
+
+            cli_routes = getattr(argument, '_hug_cli_routes', ())
+            for route in cli_routes:
+                cli(**self.where(**route).route)(argument)
 
         return method_or_class
 
@@ -73,14 +80,26 @@ class Object(http):
             for method in HTTP_METHODS:
                 handler = getattr(instance, method.lower(), None)
                 if handler:
-                    routes = getattr(handler, '_hug_routes', None)
-                    if routes:
-                        for route in routes:
+                    http_routes = getattr(handler, '_hug_http_routes', ())
+                    if http_routes:
+                        for route in http_routes:
                             http(**router.accept(method).where(**route).route)(handler)
                     else:
                         http(**router.accept(method).route)(handler)
+
+                    cli_routes = getattr(handler, '_hug_cli_routes', ())
+                    if cli_routes:
+                        for route in cli_routes:
+                            cli(**self.where(**route).route)(handler)
             return class_definition
         return decorator
+
+    def cli(self, method):
+        """Registers a method on an Object as a CLI route"""
+        routes = getattr(method, '_hug_cli_routes', [])
+        routes.append(self.route)
+        method._hug_cli_routes = routes
+        return method
 
 
 class API(object):
@@ -92,40 +111,113 @@ class API(object):
             api = hug.api.API(api)
         self.api = api
 
-    def urls(self, *kargs, **kwargs):
-        """Starts the process of building a new URL route linked to this API instance"""
+    def http(self, *args, **kwargs):
+        """Starts the process of building a new HTTP route linked to this API instance"""
         kwargs['api'] = self.api
-        return http(*kargs, **kwargs)
+        return http(*args, **kwargs)
 
-    def not_found(self, *kargs, **kwargs):
+    def urls(self, *args, **kwargs):
+        """DEPRECATED: for backwords compatibility with < hug 2.2.0. `API.http` should be used instead.
+
+           Starts the process of building a new URL HTTP route linked to this API instance
+        """
+        return self.http(*args, **kwargs)
+
+    def not_found(self, *args, **kwargs):
         """Defines the handler that should handle not found requests against this API"""
         kwargs['api'] = self.api
-        return not_found(*kargs, **kwargs)
+        return not_found(*args, **kwargs)
 
-    def static(self, *kargs, **kwargs):
+    def static(self, *args, **kwargs):
         """Define the routes to static files the API should expose"""
         kwargs['api'] = self.api
-        return static(*kargs, **kwargs)
+        return static(*args, **kwargs)
 
-    def sink(self, *kargs, **kwargs):
+    def sink(self, *args, **kwargs):
         """Define URL prefixes/handler matches where everything under the URL prefix should be handled"""
         kwargs['api'] = self.api
-        return sink(*kargs, **kwargs)
+        return sink(*args, **kwargs)
 
-    def exception(self, *kargs, **kwargs):
+    def exception(self, *args, **kwargs):
         """Defines how this API should handle the provided exceptions"""
         kwargs['api'] = self.api
-        return exception(*kargs, **kwargs)
+        return exception(*args, **kwargs)
 
-    def cli(self, *kargs, **kwargs):
+    def cli(self, *args, **kwargs):
         """Defines a CLI function that should be routed by this API"""
         kwargs['api'] = self.api
-        return cli(*kargs, **kwargs)
+        return cli(*args, **kwargs)
 
-    def object(self, *kargs, **kwargs):
+    def object(self, *args, **kwargs):
         """Registers a class based router to this API"""
         kwargs['api'] = self.api
-        return Object(*kargs, **kwargs)
+        return Object(*args, **kwargs)
+
+    def get(self, *args, **kwargs):
+        """Builds a new GET HTTP route that is registered to this API"""
+        kwargs['api'] = self.api
+        kwargs['accept'] = ('GET', )
+        return http(*args, **kwargs)
+
+    def post(self, *args, **kwargs):
+        """Builds a new POST HTTP route that is registered to this API"""
+        kwargs['api'] = self.api
+        kwargs['accept'] = ('POST', )
+        return http(*args, **kwargs)
+
+    def put(self, *args, **kwargs):
+        """Builds a new PUT HTTP route that is registered to this API"""
+        kwargs['api'] = self.api
+        kwargs['accept'] = ('PUT', )
+        return http(*args, **kwargs)
+
+    def delete(self, *args, **kwargs):
+        """Builds a new DELETE HTTP route that is registered to this API"""
+        kwargs['api'] = self.api
+        kwargs['accept'] = ('DELETE', )
+        return http(*args, **kwargs)
+
+    def connect(self, *args, **kwargs):
+        """Builds a new CONNECT HTTP route that is registered to this API"""
+        kwargs['api'] = self.api
+        kwargs['accept'] = ('CONNECT', )
+        return http(*args, **kwargs)
+
+    def head(self, *args, **kwargs):
+        """Builds a new HEAD HTTP route that is registered to this API"""
+        kwargs['api'] = self.api
+        kwargs['accept'] = ('HEAD', )
+        return http(*args, **kwargs)
+
+    def options(self, *args, **kwargs):
+        """Builds a new OPTIONS HTTP route that is registered to this API"""
+        kwargs['api'] = self.api
+        kwargs['accept'] = ('OPTIONS', )
+        return http(*args, **kwargs)
+
+    def patch(self, *args, **kwargs):
+        """Builds a new PATCH HTTP route that is registered to this API"""
+        kwargs['api'] = self.api
+        kwargs['accept'] = ('PATCH', )
+        return http(*args, **kwargs)
+
+    def trace(self, *args, **kwargs):
+        """Builds a new TRACE HTTP route that is registered to this API"""
+        kwargs['api'] = self.api
+        kwargs['accept'] = ('TRACE', )
+        return http(*args, **kwargs)
+
+    def get_post(self, *args, **kwargs):
+        """Builds a new GET or POST HTTP route that is registered to this API"""
+        kwargs['api'] = self.api
+        kwargs['accept'] = ('GET', 'POST')
+        return http(*args, **kwargs)
+
+    def put_post(self, *args, **kwargs):
+        """Builds a new PUT or POST HTTP route that is registered to this API"""
+        kwargs['api'] = self.api
+        kwargs['accept'] = ('PUT', 'POST')
+        return http(*args, **kwargs)
 
 
 for method in HTTP_METHODS: