+++ tests/test_route.py
@@ -73,6 +73,41 @@ def test_routing_class_based_method_view_with_sub_routing():
     assert hug.test.post(api, 'home').data == 'bye'
 
 
+def test_routing_class_with_cli_commands():
+    """Basic operation test"""
+    @hug.object(name='git', version='1.0.0')
+    class GIT(object):
+        """An example of command like calls via an Object"""
+
+        @hug.object.cli
+        def push(self, branch='master'):
+            return 'Pushing {}'.format(branch)
+
+        @hug.object.cli
+        def pull(self, branch='master'):
+            return 'Pulling {}'.format(branch)
+
+    assert 'token' in hug.test.cli(GIT.push, branch='token')
+    assert 'another token' in hug.test.cli(GIT.pull, branch='another token')
+
+
+def test_routing_class_based_method_view_with_cli_routing():
+    """Test creating class based routers using method mappings exposing cli endpoints"""
+    @hug.object.http_methods()
+    class EndPoint(object):
+
+        @hug.object.cli
+        def get(self):
+            return 'hi there!'
+
+        def post(self):
+            return 'bye'
+
+    assert hug.test.get(api, 'endpoint').data == 'hi there!'
+    assert hug.test.post(api, 'endpoint').data == 'bye'
+    assert hug.test.cli(EndPoint.get) == 'hi there!'
+
+
 def test_routing_instance():
     """Test to ensure its possible to route a class after it is instanciated"""
     class EndPoint(object):
@@ -97,6 +132,18 @@ class TestAPIRouter(object):
     def test_route_url(self):
         """Test to ensure you can dynamically create a URL route attached to a hug API"""
         assert self.router.urls('/hi/').route == URLRouter('/hi/', api=api).route
+
+    def test_route_http(self):
+        """Test to ensure you can dynamically create an HTTP route attached to a hug API"""
+        assert self.router.http('/hi/').route == URLRouter('/hi/', api=api).route
+
+    def test_method_routes(self):
+        """Test to ensure you can dynamically create an HTTP route attached to a hug API"""
+        for method in hug.HTTP_METHODS:
+            assert getattr(self.router, method.lower())('/hi/').route['accept'] == (method, )
+
+        assert self.router.get_post('/hi/').route['accept'] == ('GET', 'POST')
+        assert self.router.put_post('/hi/').route['accept'] == ('PUT', 'POST')
 
     def test_not_found(self):
         """Test to ensure you can dynamically create a Not Found route attached to a hug API"""